refactor: session cleanup — ensure_candidates helper, file splits, FFI tests
- Extract ensure_candidates() helper in key_handlers.rs, replacing 5 duplicated
  lazy-generate patterns (Enter, Space, Down, Up, Tab)
- Split trie_dict.rs IO methods (to_bytes, from_bytes, open, save) into
  trie_dict_io.rs (~170 lines extracted)
- Split user_history/mod.rs persistence methods into persistence.rs
  (~120 lines extracted)
- Remove unused committed_context() from FFI layer
- Add 9 FFI boundary tests for convert_to_events in api/types.rs
