+++ engine/src/api/session.rs
@@ -114,10 +114,6 @@ impl LexSession {
     fn set_abc_passthrough(&self, enabled: bool) {
         self.session.lock().unwrap().set_abc_passthrough(enabled);
     }
-
-    fn committed_context(&self) -> String {
-        self.session.lock().unwrap().committed_context()
-    }
 }
 
 impl LexSession {
