+++ engine/src/api/types.rs
@@ -127,3 +127,131 @@ pub(super) fn convert_to_events(resp: KeyResponse, has_pending_work: bool) -> Le
         events,
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::session::{CandidateAction, KeyResponse, MarkedText, SideEffects};
+
+    fn empty_response() -> KeyResponse {
+        KeyResponse {
+            consumed: false,
+            commit: None,
+            marked: None,
+            candidates: CandidateAction::Keep,
+            async_request: None,
+            side_effects: SideEffects::default(),
+        }
+    }
+
+    #[test]
+    fn test_convert_empty_response() {
+        let resp = empty_response();
+        let result = convert_to_events(resp, false);
+        assert!(!result.consumed);
+        assert!(result.events.is_empty());
+    }
+
+    #[test]
+    fn test_convert_commit_event() {
+        let mut resp = empty_response();
+        resp.consumed = true;
+        resp.commit = Some("テスト".to_string());
+        let result = convert_to_events(resp, false);
+        assert!(result.consumed);
+        assert_eq!(result.events.len(), 1);
+        assert!(matches!(&result.events[0], LexEvent::Commit { text } if text == "テスト"));
+    }
+
+    #[test]
+    fn test_convert_marked_text_event() {
+        let mut resp = empty_response();
+        resp.consumed = true;
+        resp.marked = Some(MarkedText {
+            text: "かな".to_string(),
+        });
+        let result = convert_to_events(resp, false);
+        assert_eq!(result.events.len(), 1);
+        assert!(matches!(&result.events[0], LexEvent::SetMarkedText { text } if text == "かな"));
+    }
+
+    #[test]
+    fn test_convert_clear_marked_text() {
+        let mut resp = empty_response();
+        resp.consumed = true;
+        resp.marked = Some(MarkedText {
+            text: String::new(),
+        });
+        let result = convert_to_events(resp, false);
+        // Empty marked text becomes SetMarkedText with empty string
+        assert_eq!(result.events.len(), 1);
+        assert!(matches!(&result.events[0], LexEvent::SetMarkedText { text } if text.is_empty()));
+    }
+
+    #[test]
+    fn test_convert_candidates_show() {
+        let mut resp = empty_response();
+        resp.consumed = true;
+        resp.candidates = CandidateAction::Show {
+            surfaces: vec!["候補1".to_string(), "候補2".to_string()],
+            selected: 0,
+        };
+        let result = convert_to_events(resp, false);
+        assert_eq!(result.events.len(), 1);
+        assert!(matches!(
+            &result.events[0],
+            LexEvent::ShowCandidates { surfaces, selected }
+                if surfaces.len() == 2 && *selected == 0
+        ));
+    }
+
+    #[test]
+    fn test_convert_candidates_hide() {
+        let mut resp = empty_response();
+        resp.consumed = true;
+        resp.candidates = CandidateAction::Hide;
+        let result = convert_to_events(resp, false);
+        assert_eq!(result.events.len(), 1);
+        assert!(matches!(&result.events[0], LexEvent::HideCandidates));
+    }
+
+    #[test]
+    fn test_convert_schedule_poll() {
+        let resp = empty_response();
+        let result = convert_to_events(resp, true);
+        assert_eq!(result.events.len(), 1);
+        assert!(matches!(&result.events[0], LexEvent::SchedulePoll));
+    }
+
+    #[test]
+    fn test_convert_switch_to_abc() {
+        let mut resp = empty_response();
+        resp.consumed = true;
+        resp.side_effects.switch_to_abc = true;
+        let result = convert_to_events(resp, false);
+        assert_eq!(result.events.len(), 1);
+        assert!(matches!(&result.events[0], LexEvent::SwitchToAbc));
+    }
+
+    #[test]
+    fn test_convert_multiple_events() {
+        let mut resp = empty_response();
+        resp.consumed = true;
+        resp.commit = Some("確定".to_string());
+        resp.marked = Some(MarkedText {
+            text: "次の入力".to_string(),
+        });
+        resp.candidates = CandidateAction::Show {
+            surfaces: vec!["a".to_string()],
+            selected: 0,
+        };
+        let result = convert_to_events(resp, true);
+        assert!(result.consumed);
+        // commit + marked + candidates + poll = 4
+        assert_eq!(result.events.len(), 4);
+        assert!(matches!(&result.events[0], LexEvent::Commit { .. }));
+        assert!(matches!(&result.events[1], LexEvent::SetMarkedText { .. }));
+        assert!(matches!(&result.events[2], LexEvent::ShowCandidates { .. }));
+        assert!(matches!(&result.events[3], LexEvent::SchedulePoll));
+    }
+}
